requirements.txt
pika>=1.3.0
pyyaml>=6.0
requests>=2.28.0
duckduckgo-search>=7.5.0
Pillow>=10.0.0  # Added for image resolution checking
